// frontend/src/pages/products/product.types.ts
export type Product = {
  product_id: number;
  lib: string | null;
  bar_code: string | null;
  price: number | null;
  vat_rate: number | null;
  wau_cost: number | null;
  min_stock: number | null;
  safety_stock: number | null;
  warning_stock: number | null;
  dci: string | null;
  pharma_class_id: number | null;
  type_id: number | null;
};